+++ Source/lwar/Screens/MainMenu.cs
@@ -49,7 +49,7 @@ namespace Lwar.Screens
 		/// <param name="topmost">Indicates whether the app screen is the topmost one.</param>
 		public override void Update(bool topmost)
 		{
-			_label.Area = new Rectangle(0, 0, Window.Width, Window.Height);
+			_label.Area = new Rectangle(0, 0, (int)Window.Width, (int)Window.Height);
 			_discoveryService.Update();
 		}
